+++ RELEASES.md
@@ -17,6 +17,7 @@ Version 0.10.0 (to be determined)
   `Rational::floor` and `Rational::round` now return `Rational`.
 * `Float::random_bits` and `Complex::random_bits` are now assignable
   to the number values, not to `Result` objects.
+* Deprecate `Complex::as_real_imag`.
 * Remove `Default` implementations from all types of `Float` and
   `Complex`; now the precision always has to be specified.
 * Improve top-level documentation.